Mod is now working without crashes.
Only so many units can be sent on a drakkar at a time, and that was the root
cause of the crash. This took forever to determine because it isn't fun without
crash logs or anything else useful to indicate what failed.

Units now will instead just get one more of that type in a drakkar, up to 6,
each year. 6 was chosen because 7 causes a crash, and 6 didn't appear to.

Resources will still scale with a multiplier each year.

The victory screen will update saying how much of each resource or unit you will get.